Move sources into CPU architecture package
